Remove osync installation and projects syncing
Not sure if this is temporary. But the idea is to have
everything on GitHub and delete old projects. If I
will ever decide that some project must stay private I
will think about syncing again.

On the other hand part of these setting might become useful
in the future when syncing other types of files/folders.